feat: avoid deleting avatar that is not in storage
Trying to delete github avatar would have caused errors
The delete function will only be run for avatars stored on supabase

+++ server/controllers/user/profile/put.js
@@ -30,16 +30,18 @@ const putProfile = [
       }
     }
     if (profile.background && profile.background !== req.body.background) {
-      const path = `backgrounds${profile.background.split('backgrounds')[1]}`;
-      const { success } = await deleteFile(path);
-      if (!success) {
-        return next(
-          new Sperror(
-            'Server Error',
-            'Error when deleting old background.',
-            500
-          )
-        );
+      if (profile.background.includes('supabase')) {
+        const path = `backgrounds${profile.background.split('backgrounds')[1]}`;
+        const { success } = await deleteFile(path);
+        if (!success) {
+          return next(
+            new Sperror(
+              'Server Error',
+              'Error when deleting old background.',
+              500
+            )
+          );
+        }
       }
     }
     const updatedProfile = await prisma.updateProfile(
